+++ dev/protected/extensions/imperavi-redactor-widget/changelog.md
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+1.2.23
+------
+
+- Imperavi redactor updated to 9.2.6 (samdark)
+
+1.2.22
+------
+
+- Imperavi redactor updated to 9.2.5 (samdark)
+
 1.2.21
 ------